// src/api/generated/toodlApi.zod.ts
export const todoUpdateBodySubjectMax = 255;

export const todoUpdateBodyDescriptionMax = 255;

export const todoUpdateBodyLocationMax = 255;

export const todoUpdateBodyRecurrenceRuleMax = 255;

export const todoUpdateBodyRecurrenceExceptionMax = 255;

export const todoUpdateBodyStartTimezoneMax = 255;

export const todoUpdateBodyEndTimezoneMax = 255;

export const todoUpdateBodyPositionMax = 255;



export const TodoUpdateBody = zod.object({
  "done": zod.boolean().optional(),
  "subject": zod.string().min(1).max(todoUpdateBodySubjectMax),
  "enableDeadline": zod.boolean().nullish(),
  "description": zod.string().max(todoUpdateBodyDescriptionMax).nullish(),
  "isAllDay": zod.boolean().nullish(),
  "location": zod.string().max(todoUpdateBodyLocationMax).nullish(),
  "recurrenceRule": zod.string().max(todoUpdateBodyRecurrenceRuleMax).nullish(),
  "recurrenceException": zod.string().max(todoUpdateBodyRecurrenceExceptionMax).nullish(),
  "startTimezone": zod.string().max(todoUpdateBodyStartTimezoneMax).nullish(),
  "endTimezone": zod.string().max(todoUpdateBodyEndTimezoneMax).nullish(),
  "startTime": zod.union([zod.string(),zod.iso.datetime({})]).optional(),
  "endTime": zod.union([zod.string(),zod.iso.datetime({})]).nullish(),
  "listId": zod.number().nullish(),
  "position": zod.string().max(todoUpdateBodyPositionMax).optional(),
  "categoryId": zod.number().nullish()
})